iniciar a publicação de anuncio

+++ src/main/java/br/com/motur/dealbackendservice/core/dataproviders/repository/ProviderTrimsRepository.java
@@ -1,6 +1,5 @@
 package br.com.motur.dealbackendservice.core.dataproviders.repository;
 
-import br.com.motur.dealbackendservice.core.model.ProviderModelsEntity;
 import br.com.motur.dealbackendservice.core.model.ProviderTrimsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ProviderTrimsRepository extends JpaRepository<ProviderTrimsEntity, Integer> {
+public interface ProviderTrimsRepository extends JpaRepository<ProviderTrimsEntity, Long> {
 
     @Query(value = "SELECT pt FROM ProviderTrimsEntity pt inner join fetch pt.provider p inner join fetch pt.parentProviderCatalog ppc inner join fetch pt.baseCatalog WHERE p.id = ?1")
     List<ProviderTrimsEntity> findAllByParentProviderCatalog(Integer modelId);
